requirements.txt
Flask
pystray
Pillow